fix(agt-policies): address review feedback (GLOB multiline, dedupe, docs, test layout)
Addresses reviewer feedback on the fail-closed hardening PR:

- bridge.py: GLOB->RE2 translation now prepends `(?s)` so `*`/`?` cross
  newlines, matching v4 fnmatch.translate's dotall semantics. Without it a
  `*secret*` blocked pattern failed to match multi-line field values, a
  residual fail-open (the deny silently did not fire). Adds an OPA-backed
  multiline regression test.
- snapshot.py: `_validate_budget_counter` now delegates to
  `_validate_non_negative_integer` / `_validate_non_negative_finite_number`
  instead of duplicating their two branches.
- migrate.py: document why `--write` only backs up governance files local to
  chain_root (ADR-0014 parent-deny immutability; moving parent files would
  mutate immutable deny rules and break sibling chains). Behavior unchanged.
- tests: consolidate the review-round-named files (test_pr1_*/test_pr2_*/
  test_pr3_*/test_review_*) into per-module/concern files
  (test_bridge_pattern, test_build_resolution, test_merge_deny, test_opa_runner,
  test_scope, test_discover) and fold the rest into the existing per-module
  suites (test_migrate/test_runtime/test_snapshot). Also fixes a hardcoded
  absolute opa path in one test to use `Path.home()`.

No coverage lost: 393 passed, 1 skipped (392 baseline + the new multiline
test). ruff --select E,F,W --ignore E501 clean.

Refs #3137
