+++ agent-governance-python/agt-policies/src/agt/policies/snapshot.py
@@ -51,11 +51,9 @@ def _utcnow_iso() -> str:
 
 def _validate_budget_counter(name: str, value: Any) -> None:
     if name in ("tool_call_count", "token_count"):
-        if isinstance(value, bool) or not isinstance(value, int) or value < 0:
-            raise ValueError(f"{name} must be a non-negative integer, got {value!r}")
-        return
-    if isinstance(value, bool) or not isinstance(value, (int, float)) or not math.isfinite(value) or value < 0:
-        raise ValueError(f"{name} must be a non-negative finite number, got {value!r}")
+        _validate_non_negative_integer(name, value)
+    else:
+        _validate_non_negative_finite_number(name, value)
 
 
 def _validate_non_negative_integer(name: str, value: Any) -> None:
