+++ agent-governance-python/agt-policies/tests/test_migrate.py
@@ -1,18 +1,6 @@
 # Copyright (c) Microsoft Corporation.
 # Licensed under the MIT License.
-"""Tests for :mod:`agt.cli.migrate` — the ``agt migrate v4-to-v5`` CLI.
-
-The suite covers the algorithm contract from ``plan.md`` §5 / M6.S1:
-
-- v4 artefact discovery (governance.yaml chains, GovernancePolicy calls,
-  PolicyAction.BLOCK references, CedarBackend calls, PolicyInterceptor
-  subclasses, legacy ``agent_os.policies`` imports);
-- dry-run safety (no file mutation);
-- ``--write`` side effects (manifest.yaml + Rego bundle + .v4-backup);
-- bridge-output schema validity for the GovernancePolicy materialisation
-  path;
-- Markdown structural correctness of the rendered report.
-"""
+"""Tests for cli.migrate: discovery, rendering, atomic --write with rollback."""
 
 from __future__ import annotations
 
@@ -21,16 +9,12 @@ import subprocess
 import sys
 from pathlib import Path
 from typing import Iterable
-
 import pytest
 import yaml
+from argparse import Namespace
+from typing import Any
 
 from agt.cli import migrate as migrate_mod
-
-
-# ---------------------------------------------------------------------------
-# Helpers
-# ---------------------------------------------------------------------------
 
 
 def _write_governance(path: Path, rule_name: str = "deny_dangerous_tool") -> None:
@@ -68,9 +52,95 @@ def _md_headings(text: str) -> list[str]:
     return [line for line in text.splitlines() if re.match(r"^#{1,6} ", line)]
 
 
-# ---------------------------------------------------------------------------
-# Discovery tests
-# ---------------------------------------------------------------------------
+def _iter_paths(paths: Iterable[Path]) -> list[str]:
+    return sorted(str(p) for p in paths)
+
+
+def _write_simple_governance(path: Path, rule_name: str) -> None:
+    doc = {
+        "rules": [
+            {
+                "name": rule_name,
+                "condition": {
+                    "field": "tool_call.name",
+                    "operator": "eq",
+                    "value": "rm",
+                },
+                "action": "deny",
+                "priority": 10,
+                "message": "rm is blocked",
+            }
+        ],
+        "intervention_points": {
+            "pre_tool_call": {
+                "policy_target": "$.tool_call.args",
+                "policy_target_kind": "tool_args",
+                "tool_name_from": "$.tool_call.name",
+                "policy": {"id": "agt_legacy_rules"},
+            }
+        },
+    }
+    path.write_text(yaml.safe_dump(doc, sort_keys=False), encoding="utf-8")
+
+
+def _write_dangerous_tool_governance(path: Path, rule_name: str = "deny_dangerous_tool") -> None:
+    doc = {
+        "rules": [
+            {
+                "name": rule_name,
+                "condition": {
+                    "field": "tool_call.name",
+                    "operator": "eq",
+                    "value": "rm",
+                },
+                "action": "deny",
+                "priority": 10,
+                "message": "rm is blocked",
+            }
+        ],
+        "intervention_points": {
+            "pre_tool_call": {
+                "policy_target": "$.tool_call.args",
+                "policy_target_kind": "tool_args",
+                "tool_name_from": "$.tool_call.name",
+                "policy": {"id": "agt_legacy_rules"},
+            }
+        },
+    }
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(yaml.safe_dump(doc, sort_keys=False), encoding="utf-8")
+
+
+def _write_policy_source(path: Path, source: str) -> None:
+    path.write_text(source.strip(), encoding="utf-8")
+
+
+def _write_rollback_chain_governance(path: Path, rule_name: str) -> None:
+    doc = {
+        "rules": [
+            {
+                "name": rule_name,
+                "condition": {
+                    "field": "tool_call.name",
+                    "operator": "eq",
+                    "value": "rm",
+                },
+                "action": "deny",
+                "priority": 10,
+                "message": "rm is blocked",
+            }
+        ],
+        "intervention_points": {
+            "pre_tool_call": {
+                "policy_target": "$.tool_call.args",
+                "policy_target_kind": "tool_args",
+                "tool_name_from": "$.tool_call.name",
+                "policy": {"id": "agt_legacy_rules"},
+            }
+        },
+    }
+    path.parent.mkdir(parents=True, exist_ok=True)
+    path.write_text(yaml.safe_dump(doc, sort_keys=False), encoding="utf-8")
 
 
 def test_find_governance_in_project_tree(tmp_path: Path) -> None:
@@ -229,11 +299,6 @@ from agent_os.policies.evaluator import PolicyEvaluator
     assert "PolicyEvaluator" in names
 
 
-# ---------------------------------------------------------------------------
-# Dry-run vs --write behaviour
-# ---------------------------------------------------------------------------
-
-
 def test_dry_run_does_not_touch_project_files(tmp_path: Path) -> None:
     """Default mode is dry-run and MUST be side-effect free."""
     gov = tmp_path / "governance.yaml"
@@ -331,11 +396,6 @@ policy = GovernancePolicy(
     assert (bundle / "billing_bot.rego").is_file()
 
 
-# ---------------------------------------------------------------------------
-# Report rendering
-# ---------------------------------------------------------------------------
-
-
 def test_render_report_is_valid_markdown(tmp_path: Path) -> None:
     """The Markdown report parses cleanly into headings + tables.
 
@@ -392,11 +452,6 @@ def test_dry_run_handles_empty_project(tmp_path: Path) -> None:
     assert not report.has_findings()
     text = migrate_mod.render_report(report)
     assert "No v4 artifacts detected" in text
-
-
-# ---------------------------------------------------------------------------
-# CLI entry-point smoke
-# ---------------------------------------------------------------------------
 
 
 def test_cli_module_help_lists_migrate_verb() -> None:
@@ -463,15 +518,6 @@ def test_cli_dry_run_and_write_are_mutually_exclusive(tmp_path: Path) -> None:
     assert "mutually exclusive" in proc.stderr
 
 
-# ---------------------------------------------------------------------------
-# Iteration helpers
-# ---------------------------------------------------------------------------
-
-
-def _iter_paths(paths: Iterable[Path]) -> list[str]:
-    return sorted(str(p) for p in paths)
-
-
 def test_finding_locations_carry_line_numbers(tmp_path: Path) -> None:
     """Every finding records a non-zero line number from the AST."""
     src = tmp_path / "lines.py"
@@ -493,13 +539,6 @@ def test_finding_locations_carry_line_numbers(tmp_path: Path) -> None:
     assert report.governance_policies[0].location.line >= 1
     assert report.policy_action_blocks[0].location.line >= 1
     assert report.legacy_imports[0].location.line >= 1
-
-
-# ---------------------------------------------------------------------------
-# Sanity: existing 106 tests still pass — verified by the CI matrix
-# rather than from inside this file. We only check our own module
-# imports cleanly so a broken refactor surfaces immediately.
-# ---------------------------------------------------------------------------
 
 
 def test_module_imports_cleanly() -> None:
@@ -539,3 +578,255 @@ policy = GovernancePolicy(name="x")
     # The rewrite snippet is rendered even when the bridge fails to import,
     # so users get actionable output.
     assert "GovernancePolicy" in gp.rewrite_snippet
+
+
+def test_write_migration_backs_up_only_chain_root_local_governance(
+    tmp_path: Path,
+) -> None:
+    project_root = tmp_path / "workspace"
+    chain_root = project_root / "service"
+    chain_root.mkdir(parents=True)
+    parent_governance = project_root / "governance.yaml"
+    local_governance = chain_root / "governance.yaml"
+    _write_simple_governance(parent_governance, "deny_parent")
+    _write_simple_governance(local_governance, "deny_local")
+    parent_original = parent_governance.read_text(encoding="utf-8")
+
+    finding = migrate_mod._migrate_governance_chain(
+        chain_root,
+        project_root,
+        write=True,
+    )
+
+    assert finding.error is None
+    assert parent_governance.is_file()
+    assert parent_governance.read_text(encoding="utf-8") == parent_original
+    assert not (project_root / ".governance.yaml.v4-backup").exists()
+    assert not local_governance.exists()
+    assert (chain_root / ".governance.yaml.v4-backup").is_file()
+    assert finding.backups == [chain_root / ".governance.yaml.v4-backup"]
+
+
+def test_write_chain_rolls_back_backups_when_second_move_fails(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Path,
+) -> None:
+    chain_root = tmp_path / "service"
+    first_governance = chain_root / "governance.yaml"
+    second_governance = chain_root / "governance.yml"
+    _write_dangerous_tool_governance(first_governance, "deny_first")
+    _write_dangerous_tool_governance(second_governance, "deny_second")
+    first_original = first_governance.read_text(encoding="utf-8")
+    second_original = second_governance.read_text(encoding="utf-8")
+
+    def fake_resolve_manifest(
+        root: Path,
+        action_path: Path,
+        *,
+        bundle_dir: Path,
+    ) -> dict[str, Any]:
+        assert root == tmp_path.resolve()
+        assert action_path == chain_root
+        bundle_dir.mkdir(parents=True, exist_ok=True)
+        (bundle_dir / "agt_legacy.rego").write_text(
+            "package agt.legacy\n",
+            encoding="utf-8",
+        )
+        return {
+            "agent_control_specification_version": "1.0.0-agt",
+            "metadata": {
+                "resolved_from": {
+                    "chain": [str(first_governance), str(second_governance)],
+                }
+            },
+            "extends": [],
+            "policies": {},
+            "intervention_points": {},
+        }
+
+    original_replace = Path.replace
+
+    def flaky_replace(self: Path, target: Path) -> Path:
+        if self == second_governance:
+            raise OSError("simulated second move failure")
+        return original_replace(self, target)
+
+    monkeypatch.setattr(migrate_mod, "resolve_manifest", fake_resolve_manifest)
+    monkeypatch.setattr(Path, "replace", flaky_replace)
+
+    finding = migrate_mod._migrate_governance_chain(
+        chain_root,
+        tmp_path,
+        write=True,
+    )
+
+    assert finding.error is not None
+    assert finding.error.startswith("write failed: simulated second move failure")
+    assert first_governance.read_text(encoding="utf-8") == first_original
+    assert second_governance.read_text(encoding="utf-8") == second_original
+    assert not (chain_root / ".governance.yaml.v4-backup").exists()
+    assert not (chain_root / ".governance.yml.v4-backup").exists()
+
+
+def test_migrate_governance_policy_records_mkdir_errors(tmp_path: Path) -> None:
+    _write_policy_source(
+        tmp_path / "app.py",
+        """
+from agent_os.integrations.base import GovernancePolicy
+policy = GovernancePolicy(name="strict")
+""",
+    )
+    (tmp_path / "policies").write_text("not a directory", encoding="utf-8")
+
+    report = migrate_mod.migrate_project(tmp_path, write=True)
+
+    assert report.governance_policies
+    assert any("policy migration failed for app" in err for err in report.errors)
+    assert any("policies" in err for err in report.errors)
+
+
+def test_migrate_governance_policy_records_write_errors(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Path,
+) -> None:
+    _write_policy_source(
+        tmp_path / "worker.py",
+        """
+from agent_os.integrations.base import GovernancePolicy
+policy = GovernancePolicy(name="strict")
+""",
+    )
+    manifest_path = tmp_path / "policies" / "worker.manifest.yaml"
+    original_write_text = Path.write_text
+
+    def fail_manifest_write(self: Path, *args: Any, **kwargs: Any) -> int:
+        if self == manifest_path:
+            raise OSError("simulated manifest write failure")
+        return original_write_text(self, *args, **kwargs)
+
+    monkeypatch.setattr(Path, "write_text", fail_manifest_write)
+
+    report = migrate_mod.migrate_project(tmp_path, write=True)
+
+    assert report.governance_policies
+    assert any("policy migration failed for worker" in err for err in report.errors)
+    assert any("simulated manifest write failure" in err for err in report.errors)
+
+
+def test_coerce_bridge_inputs_rejects_bool_for_numeric_fields() -> None:
+    inputs = migrate_mod._coerce_bridge_inputs(
+        {
+            "max_tokens": True,
+            "max_tool_calls": True,
+            "confidence_threshold": False,
+            "require_human_approval": True,
+        }
+    )
+
+    assert inputs.max_tokens == 0
+    assert not isinstance(inputs.max_tokens, bool)
+    assert inputs.max_tool_calls == 0
+    assert not isinstance(inputs.max_tool_calls, bool)
+    assert inputs.confidence_threshold == 0.0
+    assert isinstance(inputs.confidence_threshold, float)
+    assert inputs.require_human_approval
+
+
+def test_write_report_failure_warns_after_successful_migration(
+    monkeypatch: pytest.MonkeyPatch,
+    capsys: pytest.CaptureFixture[str],
+    tmp_path: Path,
+) -> None:
+    _write_dangerous_tool_governance(tmp_path / "governance.yaml")
+    report_path = tmp_path / "reports" / "MIGRATION.md"
+    original_write_text = Path.write_text
+
+    def fail_report_write(self: Path, *args: Any, **kwargs: Any) -> int:
+        if self == report_path:
+            raise OSError("simulated report write failure")
+        return original_write_text(self, *args, **kwargs)
+
+    monkeypatch.setattr(Path, "write_text", fail_report_write)
+
+    rc = migrate_mod.run_from_args(
+        Namespace(
+            direction="v4-to-v5",
+            project_root=str(tmp_path),
+            write=True,
+            dry_run=False,
+            write_report=str(report_path),
+            verbose=False,
+        )
+    )
+
+    captured = capsys.readouterr()
+    assert rc == 0
+    assert report_path.parent.is_dir()
+    assert "warning: failed to write report" in captured.err
+    assert str(report_path) in captured.err
+    assert "simulated report write failure" in captured.err
+
+
+def test_write_chain_rollback_restores_manifest_and_generated_bundle(
+    monkeypatch: pytest.MonkeyPatch,
+    tmp_path: Path,
+) -> None:
+    chain_root = tmp_path / "service"
+    first_governance = chain_root / "governance.yaml"
+    second_governance = chain_root / "governance.yml"
+    manifest_path = chain_root / "manifest.yaml"
+    _write_rollback_chain_governance(first_governance, "deny_first")
+    _write_rollback_chain_governance(second_governance, "deny_second")
+    manifest_path.write_text("original: manifest\n", encoding="utf-8")
+    first_original = first_governance.read_text(encoding="utf-8")
+    second_original = second_governance.read_text(encoding="utf-8")
+    manifest_original = manifest_path.read_bytes()
+
+    def fake_resolve_manifest(
+        root: Path,
+        action_path: Path,
+        *,
+        bundle_dir: Path,
+    ) -> dict[str, Any]:
+        assert root == tmp_path.resolve()
+        assert action_path == chain_root
+        bundle_dir.mkdir(parents=True, exist_ok=True)
+        (bundle_dir / "generated.rego").write_text(
+            "package agt.generated\n",
+            encoding="utf-8",
+        )
+        return {
+            "agent_control_specification_version": "1.0.0-agt",
+            "metadata": {
+                "resolved_from": {
+                    "chain": [str(first_governance), str(second_governance)],
+                }
+            },
+            "extends": [],
+            "policies": {},
+            "intervention_points": {},
+        }
+
+    original_replace = Path.replace
+
+    def flaky_replace(self: Path, target: Path) -> Path:
+        if self == second_governance:
+            raise OSError("simulated second move failure")
+        return original_replace(self, target)
+
+    monkeypatch.setattr(migrate_mod, "resolve_manifest", fake_resolve_manifest)
+    monkeypatch.setattr(Path, "replace", flaky_replace)
+
+    finding = migrate_mod._migrate_governance_chain(
+        chain_root,
+        tmp_path,
+        write=True,
+    )
+
+    assert finding.error == "write failed: simulated second move failure"
+    assert manifest_path.read_bytes() == manifest_original
+    assert first_governance.read_text(encoding="utf-8") == first_original
+    assert second_governance.read_text(encoding="utf-8") == second_original
+    assert not (chain_root / ".governance.yaml.v4-backup").exists()
+    assert not (chain_root / ".governance.yml.v4-backup").exists()
+    assert not (chain_root / "policy").exists()
